+++ packages/react/src/atomic/MultiSelect/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Use a multi-select box when a user needs to select more than one option from a list.
 
+## 2023-12-15 `1.0.1`
+
+#### 🐛 Bug fixes
+
+- Fixed bug where value was not being updated.([#203](https://github.com/TiendaNube/nimbus-design-system/pull/212) by [@vrlopess](https://github.com/vrlopess))
+
 ## 2023-10-30 `1.0.0`
 
 #### 🎉 New features